pyproject.toml
[project]
name = "zimagesr"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
  "accelerate",
  "boto3",
  "diffusers",
  "numpy",
  "pillow",
  "pyyaml",
  "safetensors",
  "torch",
  "torchvision",
  "tqdm",
  "transformers",
  "zenml",
]

[project.scripts]
zimagesr-data = "zimagesr.cli:main"
zimagesr-zenml-bootstrap = "zimagesr.pipelines.zenml_bootstrap:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/zimagesr"]
